app/requirements.txt
markdown==3.1.1
pdfkit==0.6.1
pybars3==0.9.7
python-dateutil==2.7.3
Flask==1.1.1